Warning fix
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10697 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/sound.cxx
@@ -255,12 +255,12 @@ void			openSound(const char*)
   /* initialize */
   float worldSize = BZDBCache::worldSize;
   timeSizeOfWorld = 1.414f * worldSize / SpeedOfSound;
-  for (i = 0; i < MaxEvents; i++) {
+  for (i = 0; i < (int)MaxEvents; i++) {
     events[i].samples = NULL;
     events[i].busy = false;
   }
   portUseCount = 0;
-  for (i = 0; i < FadeDuration; i += 2) {
+  for (i = 0; i < (int)FadeDuration; i += 2) {
     fadeIn[i] = fadeIn[i+1] =
 		sinf((float)(M_PI / 2.0 * (double)i / (double)(FadeDuration-2)));
     fadeOut[i] = fadeOut[i+1] = 1.0f - fadeIn[i];
